RFC: Mmoves _filter_requires_grad to skorch.utils
